quest 5 challenge

+++ app.js
@@ -104,18 +104,26 @@ app.put('/api/movies/:id', (req, res) => {
 
 app.delete('/api/users/:id', (req, res) => {
   const userId = req.params.id;
-  connection.query(
-    'DELETE FROM users WHERE id = ?',
-    [userId],
-    (err, results) => {
-      if (err) {
-        console.log(err);
-        res.status(500).send('😱 Error deleting an user');
-      } else {
-        res.status(200).send('🎉 User deleted!');
-      }
+  connection.query('DELETE FROM users WHERE id = ?', [userId], (err) => {
+    if (err) {
+      console.log(err);
+      res.status(500).send('Error deleting an user');
+    } else {
+      res.status(200).send('🎉 User deleted!');
     }
-  );
+  });
+});
+
+app.delete('/api/movies/:id', (req, res) => {
+  const movieId = req.params.id;
+  connection.query('DELETE FROM movies WHERE id = ?', [movieId], (err) => {
+    if (err) {
+      console.log(err);
+      res.status(500).send('Error deleting a movie');
+    } else {
+      res.status(200).send('🎉 Movie deleted!');
+    }
+  });
 });
 
 app.listen(port, () => {
